split webpack config and other utils to their own repos
