+++ example/package.json
@@ -14,7 +14,7 @@
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.5.0",
     "chai": "^3.5.0",
-    "eslint-config-jameslnewell": "^1.0.0-rc.7",
+    "eslint-config-jameslnewell": "^1.0.0-rc.8",
     "tradie-plugin-copy": "^0.2.0-preview.1",
     "tradie-plugin-livereload": "^0.1.2",
     "tradie-plugin-serve": "^0.2.0-preview.1"
