--- src/lib/config/extendDefaultConfig.js
@@ -1,4 +1,0 @@
-import shallowMerge from './../util/shallowMerge';
-import defaultConfig from './defaultConfig';
-
-export default userConfig => shallowMerge(defaultConfig, userConfig);
