+++ src/lib/config/readFromFile.js
