+++ src/lib/util/getRevManifestFromStats.js
@@ -1,4 +1,3 @@
-import path from 'path';
 
 export default function(stats) {
   const manifest = {};
