+++ src/lib/util/getRevManifestFromStats.test.js
