--- src/lib/webpack/mapExtensionsToRegExp.js
@@ -1,5 +1,0 @@
-import extRegex from 'ext-to-regex';
-
-export default function mapExtensionsToRegExp(extensions) {
-  return extRegex(extensions);
-}
